Make OpenIdExtension work as application library
There is apparently a corner case, that extension in WEB-INF/lib cannot
create a bean that injects something defined programmatically by an
extension that is within a server.

This limitation does not exist when bean is discovered by natural means.

Therefore the entire process changed:

OpenIdAuthenticationMechanism is disabled by means of @Alternative
annotation. When definition is found, the bean is explicitly enabled by
extension and only the definition is produced as dynamic bean.